fix: grammar compile warnings

+++ lib/lpc/program/icode.c
@@ -986,12 +986,12 @@ i_generate_if_branch (parse_node_t * node, int invert)
 }
 
 void
-i_generate_inherited_init_call (int index, short f)
+i_generate_inherited_init_call (int index, int f)
 {
   end_pushes ();
   ins_byte (F_CALL_INHERITED);
   ins_byte ((BYTE)index);
-  ins_short (f);
+  ins_short ((short)f);
   ins_byte (0);
   ins_byte (F_POP_VALUE);
 }
